% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_distance.R
\name{cluster_dist}
\alias{cluster_dist}
\title{Calculate the Czekanowski’s matrix of the clustering results.}
\usage{
cluster_dist(x, y, distfun = dist, dist_method = "average")
}
\arguments{
\item{x}{A data set or a matrix with class czek_matrix.}

\item{y}{If x is the data set, y is the cluster label.}

\item{distfun}{Specifies which distance function should be used.}

\item{dist_method}{Four linkage criteria: single, complete, average and SSD.}
}
\value{
A distance matrix.
}
\description{
Calculate the Czekanowski’s matrix of the clustering results.
}
