% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMaCzek.R
\docType{package}
\name{CzekanowskiPackage}
\alias{CzekanowskiPackage}
\alias{CzekanowskiPackage-package}
\title{CzekanowskiPackage: A package that can produce Czekanowski's diagram}
\description{
This package produce Czekanowski's diagram
}
\section{The packages functions}{

\describe{
  \item{czek_matrix}{A function that return a distance matrix where the distances is divided into classes. The return from the function is expected to be passed into the plot function. }
  \item{plot.czek_matrix}{A function that return a Czekanowski's Diagram.}
}
}

\author{
Albin Vasterlund
}
