% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.czek_matrix.R
\name{plot.czek_matrix}
\alias{plot.czek_matrix}
\title{Produce a Czekanowski's Diagram}
\usage{
\method{plot}{czek_matrix}(
  x,
  values = NULL,
  type = "symbols",
  plot_title = "Czekanowski's diagram",
  tl.cex = 1,
  tl.offset = 0.4,
  tl.srt = 90,
  pal = brewer.pal(n = 8, name = "Dark2"),
  alpha = 0.3,
  ps_power = 0.6,
  col_size = 0.8,
  cex.main = 1,
  ...
)
}
\arguments{
\item{x}{a matrix with class czek_matrix.}

\item{values}{specifies the color or the size of the symbols in the graph. The standard setting is a grey scale for a color graph and a vector with the values 2,1,0.5,0.25 and 0 for a graph with symbols.}

\item{type}{specifies if the graph should use color or symbols. The standard setting is symbols.}

\item{plot_title}{specifies the main title in the graph.}

\item{tl.cex}{Numeric, for the size of text label.}

\item{tl.offset}{Numeric, for text label.}

\item{tl.srt}{Numeric, for text label string rotation in degrees.}

\item{pal}{The colour vector represents clusters.}

\item{alpha}{Factor modifying the opacity alpha, typically in [0,1].}

\item{ps_power}{A power value to adjust point size.}

\item{col_size}{When type="col", the size of each point (maximum is 1).}

\item{cex.main}{Specify the size of the title text.}

\item{...}{specifies further parameters that can be passed on to the plot function.}
}
\description{
This is a function that can produce a Czekanowski's Diagram.
}
\examples{
# Set data ####
x<-czek_matrix(mtcars)


# Standard plot ############
plot(x)
plot.czek_matrix(x)


# Specify values ############
plot(x,values=c(1.5,1,0.75,0.25,0 ))
plot(x,values=grDevices::colorRampPalette(c("black","red","white"))(5))


# Specify type ############
plot(x,type = "symbols")
plot(x,type = "col")


# Specify the main title ############
plot(x,plot_title = "Czekanowski's Diagram of mtcars")


# Change additional settings to the plot function ############
plot(x,col.main="blue",font.main=9,cex.main=2)
}
