\name{RMaCzek-internal}
\title{Internal RMaCzek objects}
\alias{.get_parameters}
\alias{.register_seriate_ga}
\alias{.seriate_ga}
\alias{.as_matrix_czek_matrix}
\alias{.my_check_is_ga_registered}
\description{Internal RMaCzek objects.}
\details{These are not to be called by the user.}
\keyword{internal}

