% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czek_matrix.R
\name{czek_matrix}
\alias{czek_matrix}
\title{Preprocess data to produce Czekanowski's Diagram.}
\usage{
czek_matrix(
  x,
  order = "OLO",
  n_classes = 5,
  interval_breaks = NULL,
  monitor = FALSE,
  distfun = dist,
  scale_data = TRUE,
  focal_obj = NULL,
  as_dist = FALSE,
  original_diagram = FALSE,
  column_order_stat_grouping = NULL,
  dist_args = list(),
  cluster = FALSE,
  cluster_type = "exact",
  num_cluster = 3,
  sig.lvl = 0.05,
  scale_bandwidth = 0.05,
  min.size = 30,
  ...
)
}
\arguments{
\item{x}{a numeric matrix, data frame or a 'dist' object.}

\item{order}{specifies which seriation method should be applied. The standard setting is the seriation method OLO. If NA or NULL, then no seriation is done and the original ordering is saved.  The user may provide their own ordering, through a number vector of indices. Also in this case no rearrangement will be done.}

\item{n_classes}{specifies how many classes the distances should be divided into. The standard setting is 5 classes.}

\item{interval_breaks}{specifies the partition boundaries for the distances. As a standard setting, each class represents an equal amount of distances. If the interval, breaks are positive and sum up to 1, then it is assumed that they specify percentages of the distances in each interval. Otherwise if provided as a numeric vector not summing up to 1, they specify the exact boundaries for the symbols representing distance groups.}

\item{monitor}{specifies if the distribution of the distances should be visualized. The standard setting is that the distribution will not be visualized. TRUE and "cumulativ_plot" is available.}

\item{distfun}{specifies which distance function should be used. Standard setting is the dist function which uses the Euclidean distance. The first argument of the function has to be the matrix or data frame containing the data.}

\item{scale_data}{specifies if the data set should be scaled. The standard setting is that the data will be scaled.}

\item{focal_obj}{Numbers or names of objects (rows if x is a dataset and not 'dist' object) that are not to take part in the reordering procedure. These observations will be placed as last rows and columns of the output matrix. See Details.}

\item{as_dist}{If TRUE, then the distance matrix of x is returned, with object ordering, instead of the matrix with the levels assigned in place of the original distances.}

\item{original_diagram}{If TRUE, then the returned matrix corresponds as close as possible to the original method proposed by Czekanowski (1909). The levels are column specific and not matrix specific. See Details}

\item{column_order_stat_grouping}{If original_diagram is TRUE, then here one can pass the partition boundaries for the ranking in each column.}

\item{dist_args}{specifies further parameters that can be passed on to the distance function.}

\item{cluster}{If TRUE, Czekanowski’s clustering is performed.}

\item{cluster_type}{pecifies the cluster type and it can be ’exact’ or ’fuzzy’.}

\item{num_cluster}{pecifies the number of clusters.}

\item{sig.lvl}{The threshold that decides the significance of change point}

\item{scale_bandwidth}{A ratio to control the wideness of the reaching range.}

\item{min.size}{Minimum number of observations between change points.}

\item{...}{specifies further parameters that can be passed on to the seriate function in the seriation package.}
}
\value{
The function returns a matrix with class czek_matrix. The return from the function is expected to be passed to the plot function. If as_dist is passed as TRUE, then a czek_matrix_dist object is returned and this is not suitable for the plotting. As an attribute of the output the optimized criterion value is returned. However, this is a guess based on seriation::seriate()'s and seriation::criterion()'s manuals. If something else was optimized, e.g. due to user's parameters, then this will be wrong. If unable to guess, then NA saved in the attribute.
}
\description{
Preprocess data to produce Czekanowski's Diagram.
}
