% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czek_matrix.R
\name{czek_matrix}
\alias{czek_matrix}
\title{Divied a distance matrix into classes}
\usage{
czek_matrix(x, order = TRUE, n_classes = 5, interval_breaks = NULL,
  monitor = "plot", distfun = dist, scale_data = TRUE,
  control_ga = list(), ...)
}
\arguments{
\item{x}{a numeric matrix, data frame or "dist" object.}

\item{order}{the order of the observations. If TRUE the seriation function with method=SPIN_STS is used, if FALSE the present order in the distance matrix is used, if a function the function is expected to return the order the user want to use.}

\item{n_classes}{number of classes the distances should be divided into.}

\item{interval_breaks}{if "equal_width_between_classes" equal width between classes will be used, the user can specify the partition boundaries. If interval_breaks is not specified, equal amount of distances is divided into every class.}

\item{monitor}{"plot" and "cumulativ_plot" is available.}

\item{scale_data}{if the data should be scaled or not.}

\item{...}{parameters to be passed into the dist or dist_function function.}

\item{sizes_cex}{if col=FALSE the user can specify the sizes of cex}

\item{col}{if a vector with colour the classes will be expressed in colours, if TRUE standard colours will be used, if FALSE the matrix will be expressed in numbers}

\item{dist_function}{a user defined distance function. If not specified the dist function will be used.}
}
\value{
The function return a matrix with class czek_matrix. The return from the function is expected to be passed to the Czekanowski_plot function.
}
\description{
This is a function that divied the values insied a distance matrix into classes.
}
