% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet_availability}
\alias{internet_availability}
\title{Data of internet_availability}
\format{
An object of class \code{list} of length 3.
}
\usage{
internet_availability
}
\description{
Data of internet_availability
}
\keyword{datasets}
