% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maczek.R
\name{read_maczek_file}
\alias{read_maczek_file}
\title{function to load data from an mdt file (MaCzek 3.3 - http://www.antropologia.uw.edu.pl/MaCzek/maczek.html)}
\usage{
read_maczek_file(filepath)
}
\arguments{
\item{filepath}{path to file *.mdt.}
}
\value{
data.frame.
}
\description{
This software comes AS IS in the hope that it will be useful WITHOUT ANY WARRANTY, NOT even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. Please understand that there may still be bugs and errors. Use it at your own risk. We take no responsibility for any errors or omissions in this package or for any misfortune that may befall you or others as a result of its use. Please send comments and report bugs to Krzysztof Bartoszek at krzbar@protonmail.ch .
}
\author{
Piotr Jaskulski
}
