% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMaCzek.R
\docType{package}
\name{RMaCzek}
\alias{RMaCzek}
\alias{RMaCzek-package}
\title{RMaCzek: A package that can produce Czekanowski's diagram}
\description{
This package produces Czekanowski's diagram
}
\section{The packages functions}{

\describe{
  \item{czek_matrix}{A function that returns a distance matrix where the distances are divided into classes. The return from the function is expected to be passed into the plot function. }
  \item{plot.czek_matrix}{A function that returns a Czekanowski's Diagram.}
}
}

\author{
Albin Vasterlund
}
