% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seals_similarities}
\alias{seals_similarities}
\title{Data of seals_similarities}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 37 rows and 37 columns.
}
\usage{
seals_similarities
}
\description{
Data of seals_similarities
}
\keyword{datasets}
