% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skulls_distances}
\alias{skulls_distances}
\title{Data of skulls_distances}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 13 rows and 13 columns.
}
\usage{
skulls_distances
}
\description{
Data of skulls_distances
}
\keyword{datasets}
