% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Um_factor.R
\name{Um_factor}
\alias{Um_factor}
\title{Evaluate the Um function}
\usage{
Um_factor(distMatrix, order = NULL, matrix_conversion_coefficient = 1,
  inverse_um = TRUE, scale_factor = FALSE)
}
\arguments{
\item{distMatrix}{a distance matrix, preferable of class dist or matrix}

\item{order}{the order}

\item{matrix_conversion_coefficient}{a positive value added to each distance in order to avoid division by zero.}

\item{inverse_um}{if the -Um should be returned insted of Um}
}
\value{
This function return a evaluation of the Um function.
}
\description{
This function evaluate the Um function.
}
\examples{
# Make a distance matrix
distMatrix<-dist(x = mtcars,diag = TRUE,upper = TRUE)

# Evaluate the order
Um_v3(distMatrix)
}
