% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urns}
\alias{urns}
\title{Data of urns}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 15 rows and 9 columns.
}
\usage{
urns
}
\description{
Data of urns
}
\keyword{datasets}
